Move subsystem.py and its test into pants.option.
Subsystem is now the only kind of Optionable, so there
is no justification for a separate top-level directory for the
Subsystem concept.

[ci skip-rust]

[ci skip-build-wheels]
