+++ src/python/pants/auth/basic_auth.py
@@ -9,7 +9,7 @@ import requests
 import www_authenticate
 
 from pants.auth.cookies import Cookies
-from pants.subsystem.subsystem import Subsystem
+from pants.option.subsystem import Subsystem
 from pants.version import VERSION
