+++ src/python/pants/backend/project_info/source_file_validator.py
@@ -13,7 +13,7 @@ from pants.engine.console import Console
 from pants.engine.fs import Digest, DigestContents, SourcesSnapshot
 from pants.engine.goal import Goal, GoalSubsystem
 from pants.engine.rules import Get, collect_rules, goal_rule
-from pants.subsystem.subsystem import Subsystem
+from pants.option.subsystem import Subsystem
 from pants.util.frozendict import FrozenDict
 from pants.util.memo import memoized_method
