+++ src/python/pants/backend/python/dependency_inference/rules.py
@@ -27,7 +27,7 @@ from pants.engine.target import (
 )
 from pants.engine.unions import UnionRule
 from pants.option.global_options import OwnersNotFoundBehavior
-from pants.subsystem.subsystem import Subsystem
+from pants.option.subsystem import Subsystem
 
 
 class PythonInference(Subsystem):
