+++ src/python/pants/core/util_rules/external_tool.py
@@ -10,7 +10,7 @@ from pants.core.util_rules.archive import ExtractedDigest, MaybeExtractable
 from pants.engine.fs import Digest, DownloadFile
 from pants.engine.platform import Platform
 from pants.engine.rules import Get, RootRule, collect_rules, rule
-from pants.subsystem.subsystem import Subsystem
+from pants.option.subsystem import Subsystem
 from pants.util.meta import classproperty
