+++ src/python/pants/help/help_info_extracter_test.py
@@ -14,7 +14,7 @@ from pants.option.options import Options
 from pants.option.parser import Parser
 from pants.option.ranked_value import Rank, RankedValue
 from pants.option.scope import GLOBAL_SCOPE
-from pants.subsystem.subsystem import Subsystem
+from pants.option.subsystem import Subsystem
 
 
 class LogLevel(Enum):
