+++ src/python/pants/init/util.py
@@ -7,7 +7,7 @@ from typing import cast
 from pants.fs.fs import safe_filename_from_path
 from pants.init.options_initializer import BuildConfigInitializer
 from pants.option.option_value_container import OptionValueContainer
-from pants.subsystem.subsystem import Subsystem
+from pants.option.subsystem import Subsystem
 from pants.util.dirutil import absolute_symlink, safe_mkdir, safe_rmtree
