+++ src/python/pants/source/source_root.py
@@ -11,7 +11,7 @@ from typing import Iterable, Optional, Set, Tuple, Union
 from pants.engine.collection import DeduplicatedCollection
 from pants.engine.fs import PathGlobs, Snapshot
 from pants.engine.rules import Get, MultiGet, RootRule, collect_rules, rule
-from pants.subsystem.subsystem import Subsystem
+from pants.option.subsystem import Subsystem
 from pants.util.memo import memoized_method
 
 logger = logging.getLogger(__name__)
