--- src/python/pants/subsystem/BUILD
@@ -1,6 +1,0 @@
-# Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
-# Licensed under the Apache License, Version 2.0 (see LICENSE).
-
-python_library()
-
-python_tests(name='tests')
