+++ src/python/pants/testutil/test_base.py
@@ -23,8 +23,8 @@ from pants.init.engine_initializer import EngineInitializer
 from pants.init.util import clean_global_runtime_state
 from pants.option.global_options import ExecutionOptions
 from pants.option.options_bootstrapper import OptionsBootstrapper
+from pants.option.subsystem import Subsystem
 from pants.source import source_root
-from pants.subsystem.subsystem import Subsystem
 from pants.testutil.engine.util import Params, init_native
 from pants.util.collections import assert_single_element
 from pants.util.contextutil import temporary_dir
